separate into multi_file_list and multi_file_reader
