+++ CMakeLists.txt
@@ -36,6 +36,7 @@ set(EXTENSION_SOURCES
     src/common/iceberg.cpp
     src/common/api_utils.cpp
     src/iceberg_functions/iceberg_multi_file_reader.cpp
+    src/iceberg_functions/iceberg_multi_file_list.cpp
     src/iceberg_functions/iceberg_snapshots.cpp
     src/iceberg_functions/iceberg_scan.cpp
     src/iceberg_functions/iceberg_metadata.cpp
